requirements.txt
asgiref==3.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-environ==0.8.1
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
importlib-metadata==4.11.2
Markdown==3.3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
